Set proxy-url in kubeconfig (temporarily)
SOCKS proxy is incompatible with awscli. This normally doesn't
matter the way this module is used, but when a role is specified
then awscli needs to call out and fails at the socks proxy.

By moving the proxy config into the kubeconfig, kubectl uses the
proxy and awscli doesn't, fixing this issue.